vscode extension: Share the properties view adaptation with the web extension
